Update patch set 1
Patch Set 1:

> Hi William
 > Can you explain what you did here? Was this change fully tested?

The base_sdc Dockerfiles were made to be arm and x86_64 compatible, the changes were only tested by running the containers and checking for obvious errors.
I couldn't find any tests for the images, so if there are some that the containers can be run against then that would be nice.

There were two main Dockerfiles I wasn't certain of; with Kibana and Elasticsearch I wasn't certain of how Chef was to be installed in the image, the recommended way of using omnitruck isn't arm-compatible. In the other Dockerfiles Chef was gem installed, so I'm not sure how you'd want it incorporated.

This submission is also an attempt at image-minimization for https://jira.onap.org/browse/INT-550

Patch-set: 1
Label: Code-Review=0