Update patch set 5
Patch Set 5:

I think the idea is trying to follow what the base image are trying to do, if they set labels, our derived images should do the same. So yes there may be a broader discussion on what labels we want, how we want them etc.

On using the Change-ID, I'm not fully sure, we can add it as well if needed but my thinking was to stick with git attributes to stay portable on other envs.

Patch-set: 5